requirements.txt
chainlit==1.0.505
huggingface_hub==0.16.4
neo4j==5.15.0
requests==2.31.0
beautifulsoup4==4.12.2
markdownify==0.11.6
sentence-transformers==2.2.2
groq==0.4.2
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
tiktoken==0.5.2
faiss-cpu==1.7.4